try to add some datum testing stuff
I'm trying to make a thing that can autogenerate a pgtest file from
the interesting datums, which should then be auto rewritten from
postgres. Figuring out where to put this file and what is the thing
that should do the creation of it and then something needs to test it
is the current problem. Should pgtest/src/main.rs generate it? Should
it be pgrepr/src/value.rs's test module? I have no idea. This is
annoying enough that I don't care right now.