coord: prevent compaction of indexes for in-progress transactions
The previous commit chooses a timestamp that will be correct for a
transaction, but does nothing to guarantee the availability of that
timestamp for the duration of the transaction.

Teach the coordinator to merge compaction information from both the upper
updates and in-progress transactions, and not compact past any of those.